enlarge batch size and move dtype conversion to corpus loading func

+++ src/corpus/dataset.py
@@ -15,9 +15,9 @@ def load_corpus(filename, vocab_ent, vocab_rel):
         if config.USE_TINY_CORPUS_NUM > 0 and i / config.USE_TINY_CORPUS_NUM >= 1:
             break
 
-        heads.append(np.array([vocab_ent(h)]))
-        rels.append(np.array([vocab_rel(r)]))
-        tails.append(np.array([vocab_ent(t)]))
+        heads.append(np.array([vocab_ent(h)], dtype=np.int32))
+        rels.append(np.array([vocab_rel(r)], dtype=np.int32))
+        tails.append(np.array([vocab_ent(t)], dtype=np.int32))
 
     return chainer.datasets.TupleDataset(heads, rels, tails)
